Move requests to new package
