+++ app/src/main/java/com/vaultionizer/vaultapp/data/model/rest/request/CreateSpaceRequest.kt
@@ -1,4 +1,4 @@
-package com.vaultionizer.vaultapp.data.model.rest.space
+package com.vaultionizer.vaultapp.data.model.rest.request
 
 data class CreateSpaceRequest(
     val authKey: String,
