+++ app/src/main/java/com/vaultionizer/vaultapp/data/model/rest/request/CreateUserRequest.kt
@@ -1,4 +1,4 @@
-package com.vaultionizer.vaultapp.data.model.rest.user
+package com.vaultionizer.vaultapp.data.model.rest.request
 
 data class CreateUserRequest(
     val key: String,
