+++ app/src/main/java/com/vaultionizer/vaultapp/data/model/rest/request/GetAuthKeyRequest.kt
@@ -1,0 +1,5 @@
+package com.vaultionizer.vaultapp.data.model.rest.request
+
+data class GetAuthKeyRequest(
+    val spaceID: Long
+)
