+++ app/src/main/java/com/vaultionizer/vaultapp/data/model/rest/request/LoginUserRequest.kt
@@ -1,4 +1,4 @@
-package com.vaultionizer.vaultapp.data.model.rest.user
+package com.vaultionizer.vaultapp.data.model.rest.request
 
 data class LoginUserRequest(
     val username: String,
