+++ app/src/main/java/com/vaultionizer/vaultapp/data/model/rest/request/UploadReferenceFileRequest.kt
@@ -1,4 +1,4 @@
-package com.vaultionizer.vaultapp.data.model.rest.rf
+package com.vaultionizer.vaultapp.data.model.rest.request
 
 data class UploadReferenceFileRequest(
     val content: String,
