--- app/src/main/java/com/vaultionizer/vaultapp/data/model/rest/space/GetAuthKeyRequest.kt
@@ -1,5 +1,0 @@
-package com.vaultionizer.vaultapp.data.model.rest.space
-
-data class GetAuthKeyRequest(
-    val spaceID: Long
-)
