+++ app/src/main/java/com/vaultionizer/vaultapp/service/UserService.kt
@@ -1,8 +1,8 @@
 package com.vaultionizer.vaultapp.service
 
 import com.vaultionizer.vaultapp.data.model.rest.result.ApiResult
-import com.vaultionizer.vaultapp.data.model.rest.user.CreateUserRequest
-import com.vaultionizer.vaultapp.data.model.rest.user.LoginUserRequest
+import com.vaultionizer.vaultapp.data.model.rest.request.CreateUserRequest
+import com.vaultionizer.vaultapp.data.model.rest.request.LoginUserRequest
 import com.vaultionizer.vaultapp.data.model.rest.user.NetworkUserAuthPair
 import retrofit2.http.Body
 import retrofit2.http.POST
